goquant_trade_simulator/requirements.txt
websocket-client>=1.6.2
PyQt5>=5.15.10
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.8.0
scipy>=1.12.0 